Move component helper docs to @ember/helper
